Updates spec_helper
Adds a `config.mock_with` block to suppress deprecation warnings when using `stub` in specs. When calling `solve` on insoluble or already solved grids, an appropriate message is printed to the console in addition to passing back a value. I use `stub` in the specs to prevent this message from being written while running tests.